Moved loading transactions list for TransactionProvider component
