--- src/TransactionContext.ts
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const TransactionContext = createContext([])
