+++ src/components/TransactionsTable/index.tsx
@@ -1,23 +1,9 @@
-import { useEffect, useState } from "react";
-import { api } from "../../services/api";
+import { useContext } from "react";
+import { TransactionContext } from "../../TransactionContext";
 import { Container } from "./styles";
 
-interface Transaction {
-  id: number,
-  title: string,
-  value: number,
-  type: string,
-  category: string,
-  createdDate: Date
-}
-
 export function TransactionsTable() {
-  const [transactions, setTransactions] = useState<Transaction[]>([])
-
-  useEffect(() => {
-    api.get('transactions')
-      .then(response => setTransactions(response.data.transactions))
-  }, []);
+  const transactions = useContext(TransactionContext);
 
   return (
     <Container>
